requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Faker==18.9.0
idna==3.7
jsonschema==4.17.3
jsonschema-spec==0.1.6
lazy-object-proxy==1.10.0
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.5
pathable==0.4.3
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
PyYAML==5.3.1
requests==2.32.3
rfc3339-validator==0.1.4
setuptools==70.0.0
six==1.16.0
urllib3==2.2.1
wheel==0.43.0
